Write tests with parser
